Add files via upload
this is a simple cuda program calculating vector add for 2 dimensions on GPU device
I added two two-dimensional matrices A, B on the device GPU. 
After the device matrix addition kernel function is invoked, and the addition result is transferred back to the CPU. 
The program will also compute the  sum matrix of matrices A and B using the CPU.  
Then the program compares the device-computed result with the CPU-computed result. 
If it matches, it prints out Test PASSED to the screen before exiting.